Source Native Support (#1231)
* Add `try-catch` fullJS test snippet

* Source Native support

- format code

* Add Source Native unit tests

- shifted CodeSnippetTestCase to testing utils

* Enable interpreter for Source Native
